+++ src/main/java/raven/datetime/component/date/PanelDate.java
@@ -24,7 +24,7 @@ public class PanelDate extends JPanel {
     private void init() {
         putClientProperty(FlatClientProperties.STYLE, "" +
                 "background:null");
-        setLayout(new MigLayout("wrap 7,insets 3,gap 1,al center center", "fill", "[fill]10[fill][fill]"));
+        setLayout(new MigLayout("novisualpadding,wrap 7,insets 3,gap 0,al center center", "fill", "[fill]10[fill][fill]"));
         load();
     }