+++ src/main/java/raven/datetime/component/date/PanelYear.java
@@ -23,7 +23,7 @@ public class PanelYear extends JPanel {
     private void init() {
         putClientProperty(FlatClientProperties.STYLE, "" +
                 "background:null");
-        setLayout(new MigLayout("wrap 4,insets 3,fillx,gap 3,al center center", "fill,sg main", "fill"));
+        setLayout(new MigLayout("novisualpadding,wrap 4,insets 0,fillx,gap 0,al center center", "fill,sg main", "fill"));
         int count = YEAR_CELL;
         for (int i = 0; i < count; i++) {
             final int y = getStartYear(year) + i;